refactor: 값 타입 필드 일부를 엔티티로 이전
