+++ src/main/java/com/stcom/smartmealtable/domain/Address/Address.java
@@ -1,8 +1,6 @@
 package com.stcom.smartmealtable.domain.Address;
 
 import jakarta.persistence.Embeddable;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,14 +16,9 @@ public class Address {
 
     private String detailAddress;
 
-    private String alias;
-
     private Double latitude;
 
     private Double longitude;
-
-    @Enumerated(EnumType.STRING)
-    private AddressType type;
 
     @Builder
     public Address(String lotNumberAddress, String roadAddress, String detailAddress, String alias, Double latitude,
@@ -33,10 +26,8 @@ public class Address {
         this.lotNumberAddress = lotNumberAddress;
         this.roadAddress = roadAddress;
         this.detailAddress = detailAddress;
-        this.alias = alias;
         this.latitude = latitude;
         this.longitude = longitude;
-        this.type = type;
     }
 
 
@@ -49,7 +40,5 @@ public class Address {
         this.longitude = longitude;
     }
 
-    public void changeAddressType(AddressType newType) {
-        this.type = newType;
-    }
+
 }
