+++ src/main/java/com/stcom/smartmealtable/domain/Address/AddressEntity.java
@@ -3,10 +3,13 @@ package com.stcom.smartmealtable.domain.Address;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -26,6 +29,18 @@ public class AddressEntity {
 
     private boolean primary = false;
 
+    @Enumerated(EnumType.STRING)
+    private AddressType type;
+
+    private String alias;
+
+    @Builder
+    public AddressEntity(Address address, AddressType type, String alias) {
+        this.address = address;
+        this.type = type;
+        this.alias = alias;
+    }
+
     public AddressEntity(Address address) {
         this.address = address;
     }
@@ -40,6 +55,14 @@ public class AddressEntity {
 
     public boolean isPrimaryAddress() {
         return primary;
+    }
+
+    public void changeAddressType(AddressType newType) {
+        this.type = newType;
+    }
+
+    public void changeAlias(String newAlias) {
+        this.alias = newAlias;
     }
